Brand new board
It is a complete rewrite of the board.
The board is now able to group tasks by the owner
The board is now able to layout the tasks onto on one another, this
makes them take less space, and furthmore, if there are more than 5
tasks in a group they will lay out in two columns.
there is a ton of more changes...